requirements.txt
tweepy
python-dotenv
requests
bs4
selenium